pyproject.toml
[project]
name = "arquisoft"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "gunicorn>=23.0.0",
    "psycopg[binary,pool]>=3.2.3",
]

[tool.uv]
dev-dependencies = [
    "locust>=2.32.0",
    "ruff>=0.7.0",
]

[tool.locust]
locustfile = "test/load"
processes = 2
host = "https://localhost:8080"
